5th Commit
Tested some features:

I saw a tutorial about Jumping Higher if the player holds a Key.
Worked! BUT, The Script Works ON THE WHOLE PROJECT.

But fortunately, I saved the Original CODE and The CODE from the tutorial on Notepad. SO, I just need to copy and paste, AND CREATE ANOTHER SCRIPT to make my stupid moves XD.

Commit Made:
14.07.2023 - 10:55 pm (22.55)